Explicitly cast i64 tokens to i32

+++ ai_edge_torch/convert/fx_passes/test/test_build_aten_composite_pass.py
@@ -130,14 +130,14 @@ class TestBuildAtenCompositePass(unittest.TestCase):
 
   def test_embedding_lookup_layer(self):
     stablehlo = _export_to_stablehlo_with_composite(
-        torch.nn.Embedding(10, 10), (torch.full((1, 10), 0, dtype=torch.int),)
+        torch.nn.Embedding(10, 10), (torch.full((1, 10), 0, dtype=torch.long),)
     )
     self.assertEqual(stablehlo.count('stablehlo.composite "odml.embedding_lookup"'), 1)
 
   def test_embedding_lookup_op(self):
     stablehlo = _export_to_stablehlo_with_composite(
         lambda *x: torch.ops.aten.embedding.default(*x),
-        (torch.rand(10, 10), torch.full((1, 10), 0, dtype=torch.int)),
+        (torch.rand(10, 10), torch.full((1, 10), 0, dtype=torch.long)),
     )
     self.assertEqual(stablehlo.count('stablehlo.composite "odml.embedding_lookup"'), 1)
 
@@ -145,7 +145,7 @@ class TestBuildAtenCompositePass(unittest.TestCase):
     stablehlo = _export_to_stablehlo_with_composite(
         lambda *x: torch.nn.functional.embedding(*x),
         (
-            torch.full((1, 10), 0, dtype=torch.int),
+            torch.full((1, 10), 0, dtype=torch.long),
             torch.rand(10, 10),
         ),
     )
